+++ app/Http/Controllers/ExpenseController.php
@@ -15,6 +15,9 @@ class ExpenseController extends Controller
     }
     public function index()
     {
+        if (!userHasPermission('expense-index')) {
+            return view('404');
+        }
         $item = $this->baseService->Index();
         $categories = ExpenseCategory::get();
         $columns = Expense::$columns;
@@ -25,6 +28,9 @@ class ExpenseController extends Controller
     }
     public function trash()
     {
+        if (!userHasPermission('expense-advance')) {
+            return view('404');
+        }
         $item = $this->baseService->Trash();
         $columns = Expense::$columns;
         if (request()->ajax()) {
@@ -35,6 +41,9 @@ class ExpenseController extends Controller
 
     public function store(Request $request)
     {
+        if (!userHasPermission('expense-store')) {
+            return view('404');
+        }
         if ($request->id == null) {
             $request->validate([
                 'amount' => 'required',
@@ -46,6 +55,9 @@ class ExpenseController extends Controller
     }
     function edit($id)
     {
+        if (!userHasPermission('expense-update')) {
+            return view('404');
+        }
         $expense = Expense::findOrFail($id);
         $item = $this->baseService->Index();
         $categories = ExpenseCategory::get();
@@ -57,18 +69,35 @@ class ExpenseController extends Controller
     }
     function delete($id)
     {
+        if (!userHasPermission('expense-delete')) {
+            return view('404');
+        }
         Expense::findOrFail($id)->delete();
         return redirect()->route('expense.index')->with('success', 'Expense Deleted Successfully');
     }
 
     function restore($id)
     {
+        if (!userHasPermission('expense-advance')) {
+            return view('404');
+        }
         Expense::withTrashed()->findOrFail($id)->restore();
         return redirect()->route('expense.index')->with('success', 'Expense Restore Successfully');
     }
     function pdelete($id)
     {
+        if (!userHasPermission('expense-advance')) {
+            return view('404');
+        }
         Expense::withTrashed()->findOrFail($id)->forceDelete();
         return redirect()->route('expense.trash')->with('success', 'Expense Permanently Deleted Successfully');
+    }
+    function receipt($id)
+    {
+        if (!userHasPermission('expense-advance')) {
+            return view('404');
+        }
+        $expense = Expense::findOrFail($id);
+        return view('pages.expense.receipt', compact('expense'));
     }
 }
