+++ app/Http/Controllers/SalarySubmitController.php
@@ -15,6 +15,9 @@ class SalarySubmitController extends Controller
     }
     public function index()
     {
+        if (!userHasPermission('payrole-index')) {
+            return view('404');
+        }
         $item = $this->baseService->Index();
         $columns = MonthlySalary::$columns;
         if (request()->ajax()) {
@@ -24,6 +27,9 @@ class SalarySubmitController extends Controller
     }
     public function create()
     {
+        if (!userHasPermission('payrole-store')) {
+            return view('404');
+        }
         $employees = Employee::select('name', 'id', 'contact')->get();
 
         return view('pages.salary_submit.create', compact('employees'));
@@ -31,24 +37,36 @@ class SalarySubmitController extends Controller
 
     public function store(Request $request)
     {
+        if (!userHasPermission('payrole-store')) {
+            return view('404');
+        }
         $data = $request->all();
         $message = $this->baseService->create($data);
         return redirect()->route('salary-submit.index')->with($message);
     }
     function edit($id)
     {
+        if (!userHasPermission('payrole-update')) {
+            return view('404');
+        }
         $salary = MonthlySalary::findOrFail($id);
         $employees = Employee::select('name', 'id', 'contact')->get();
         return view('pages.Salary_submit.edit', compact('salary', 'employees'));
     }
     function update(Request $request, $id)
     {
+        if (!userHasPermission('payrole-update')) {
+            return view('404');
+        }
         $data = $request->all();
         $message = $this->baseService->update($data, $id);
         return redirect()->route('salary-submit.index')->with($message);
     }
     function delete($id)
     {
+        if (!userHasPermission('payrole-delete')) {
+            return view('404');
+        }
         MonthlySalary::findOrFail($id)->delete();
         return redirect()->route('salary-submit.index')->with('success', 'Monthly Salary Deleted Successfully');
     }
